Correcciones y CRUD de los maestros Cargo y Competencia

+++ Bizapps-evaluacion/src/interfacedao/maestros/ICompetenciaDAO.java
@@ -6,7 +6,7 @@ import modelos.Competencia;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-public interface ICompetenciaDAO extends JpaRepository<Competencia, Long> {
+public interface ICompetenciaDAO extends JpaRepository<Competencia, Integer> {
 	
 	
 	 @Query("select a from Competencia a where a.nivel = 'RECTORAS'")
